+++ src/p_mobj.c
@@ -293,8 +293,9 @@ void P_XYMovement (mobj_t* mo)
 
   // no friction for missiles or skulls ever, no friction when airborne
   if (mo->flags & (MF_MISSILE | MF_SKULLFLY)
-      // [Nugget] Do apply friction if airborne with the flight cheat on
-      || (!(player && player->cheats & CF_FLY) && (mo->z > mo->floorz)))
+      // [Nugget] Do apply friction if airborne with noclip or flight cheat enabled
+      || (mo->z > mo->floorz
+          && !(casual_play && player && (mo->flags & MF_NOCLIP || player->cheats & CF_FLY))))
     return;
 
   // killough 8/11/98: add bouncers